requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
sentence-transformers
pinecone-client
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
PyPDF2
googletrans
SpeechRecognition
pydub
gTTS
numpy
pandas
langchain
langchain-google-genai
torch
transformers
aiofiles
Pillow
bcrypt